+++ packages/moonstone/TooltipDecorator/TooltipDecorator.js
@@ -20,6 +20,26 @@ import {Tooltip, TooltipBase} from './Tooltip';
 let currentTooltip; // needed to know whether or not we should stop a showing job when unmounting
 
 /**
+ * Default config for {@link moonstone/TooltipDecorator.TooltipDecorator}
+ *
+ * @memberof moonstone/TooltipDecorator.TooltipDecorator
+ * @hocconfig
+ */
+const defaultConfig = {
+	/**
+	 * The name of the property which will receive the tooltip node. By default, `TooltipDecorator`
+	 * will add a new child to the wrapped component, following any other children passed in. If
+	 * a component needs to, it can specify another property to receive the tooltip and the
+	 * `children` property will not be modified.
+	 *
+	 * @type {String}
+	 * @default 'children'
+	 * @memberof moonstone/TooltipDecorator.TooltipDecorator.defaultConfig
+	 */
+	tooltipDestinationProp: 'children'
+};
+
+/**
  * {@link moonstone/TooltipDecorator.TooltipDecorator} is a Higher-order Component which
  * positions {@link moonstone/TooltipDecorator.Tooltip} in relation to the
  * Wrapped component.
@@ -34,12 +54,13 @@ let currentTooltip; // needed to know whether or not we should stop a showing jo
  * @hoc
  * @public
  */
-const TooltipDecorator = hoc((config, Wrapped) => {
+const TooltipDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 	const forwardBlur = forward('onBlur');
 	const forwardFocus = forward('onFocus');
 	const forwardMouseOver = forward('onMouseOver');
 	const forwardMouseOut = forward('onMouseOut');
+	const tooltipDestinationProp = config.tooltipDestinationProp;
 
 	return class extends React.Component {
 		static displayName = 'TooltipDecorator'
@@ -349,6 +370,33 @@ const TooltipDecorator = hoc((config, Wrapped) => {
 			delete rest.tooltipDelay;
 			delete rest.tooltipPosition;
 
+			const renderedTooltip = (
+				<FloatingLayer open={this.state.showing} scrimType="none" key="tooltipFloatingLayer">
+					<Tooltip
+						aria-live="off"
+						role="alert"
+						{...tooltipProps}
+						arrowAnchor={this.state.arrowAnchor}
+						casing={tooltipCasing}
+						direction={this.state.tooltipDirection}
+						position={this.state.position}
+						preserveCase={tooltipPreserveCase}
+						tooltipRef={this.getTooltipRef}
+						width={tooltipWidth}
+					>
+						{tooltipText}
+					</Tooltip>
+				</FloatingLayer>
+			);
+
+			let internalTooltip;
+
+			if (tooltipDestinationProp !== 'children') {
+				rest[tooltipDestinationProp] = renderedTooltip;
+			} else {
+				internalTooltip = [children, renderedTooltip];
+			}
+
 			return (
 				<Wrapped
 					{...rest}
@@ -357,23 +405,7 @@ const TooltipDecorator = hoc((config, Wrapped) => {
 					onMouseOut={this.handleMouseOut}
 					onMouseOver={this.handleMouseOver}
 				>
-					{children}
-					<FloatingLayer open={this.state.showing} scrimType="none">
-						<Tooltip
-							aria-live="off"
-							role="alert"
-							{...tooltipProps}
-							arrowAnchor={this.state.arrowAnchor}
-							casing={tooltipCasing}
-							direction={this.state.tooltipDirection}
-							position={this.state.position}
-							preserveCase={tooltipPreserveCase}
-							tooltipRef={this.getTooltipRef}
-							width={tooltipWidth}
-						>
-							{tooltipText}
-						</Tooltip>
-					</FloatingLayer>
+					{internalTooltip || children}
 				</Wrapped>
 			);
 		}